Remove the duplicate definition of timeouts
After the last changes, the timeouts were defined both on the connection level and on the request level.
This commit cancel the previous update and keep the definition on the request level. The timeouts value have been updated to 5s for the open_timeout and 30s for the read_timeout.

https://creativegroupdev.atlassian.net/browse/CT-628